+++ benchmark/c/resource_utils.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 namespace benchmark::utils {
 
@@ -11,9 +12,9 @@ size_t GetPeakWorkingSetSizeInBytes();
 
 #ifdef _WIN32
 struct GpuMemoryInfo {
-  size_t dedicated;  // VRAM (DXGI_MEMORY_SEGMENT_GROUP_LOCAL)
-  size_t shared;     // System RAM used by GPU (DXGI_MEMORY_SEGMENT_GROUP_NON_LOCAL)
-  size_t Total() const { return dedicated + shared; }
+  uint64_t dedicated;  // VRAM (DXGI_MEMORY_SEGMENT_GROUP_LOCAL)
+  uint64_t shared;     // System RAM used by GPU (DXGI_MEMORY_SEGMENT_GROUP_NON_LOCAL)
+  uint64_t Total() const { return dedicated + shared; }
 };
 
 GpuMemoryInfo GetGpuMemoryUsage();
